requirements.txt
# Core dependencies
aiofiles==24.1.0
annotated-types==0.6.0
anyio==4.7.0
certifi==2025.4.26
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
jiter==0.11.0
mpmath==1.3.0
numpy==2.1.3
openai==2.5.0
pandas==2.2.3
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sniffio==1.3.0
sympy==1.13.3
tenacity==9.0.0
tqdm==4.67.1
tree-sitter==0.25.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
requests==2.32.3

# Additional dependencies
pydantic-settings==2.6.1